+++ test/integration/test-amp-carousel.js
@@ -1,5 +1,5 @@
 /**
- * Copyright 2015 The AMP HTML Authors. All Rights Reserved.
+ * Copyright 2018 The AMP HTML Authors. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,249 +14,351 @@
  * limitations under the License.
  */
 
-import {AmpEvents} from '../../src/amp-events';
-import {
-  createFixtureIframe,
-  expectBodyToBecomeVisible,
-} from '../../testing/iframe';
+const config = describe.configure().ifNewChrome();
+config.run('amp-carousel', function() {
+  this.timeout(10000);
 
-describe.skip('integration amp-carousel', () => {
+  const extensions = ['amp-carousel'];
 
-  let fixture;
-  beforeEach(() => {
-    return createFixtureIframe('test/fixtures/carousels.html', 1000)
-        .then(f => {
-          fixture = f;
-        });
-  });
+  const carouselSingleImage = `
+  <amp-carousel width=400 height=300 id="carousel-1">
+    <amp-img src="https://lh3.googleusercontent.com/pSECrJ82R7-AqeBCOEPGPM9iG9OEIQ_QXcbubWIOdkY=w400-h300-no" layout=fill></amp-img>
+  </amp-carousel>
+  `;
 
-  it('should show the body in carousel test', () => {
-    return expectBodyToBecomeVisible(
-        fixture.win, window.ampTestRuntimeConfig.mochaTimeout);
-  });
-
-  it('should be present', () => {
-    expect(fixture.doc.querySelectorAll('amp-carousel'))
-        .to.have.length.above(0);
-    return fixture.awaitEvent(AmpEvents.LOAD_START, 1).then(() => {
-      expect(fixture.doc.querySelectorAll('amp-carousel'))
-          .to.have.length.above(0);
-    });
-  });
-
-  describe('when amp-mode-mouse class is on body', () => {
+  describes.integration('type=carousel with single image', {
+    body: carouselSingleImage,
+    extensions,
+  }, env => {
+    let document;
 
     beforeEach(() => {
-      fixture.doc.body.classList.add('amp-mode-mouse');
+      document = env.win.document;
     });
 
-    it('should only have the next button enabled ' +
-       'when on first item', () => {
-      return fixture.awaitEvent(AmpEvents.LOAD_START, 1).then(() => {
-        fixture.doc.body.classList.add('amp-mode-mouse');
-        const amp = fixture.doc.querySelector('#carousel-1');
-        const prevBtn = amp.querySelector('.amp-carousel-button-prev');
-        const nextBtn = amp.querySelector('.amp-carousel-button-next');
-        expect(prevBtn).to.be.hidden;
-        expect(nextBtn).to.be.visible;
-      });
+    it('should be present', () => {
+      expect(document.querySelectorAll('amp-carousel')).to.have.length.above(0);
     });
 
-    it('should not be able to go past the first or last item', () => {
-      return fixture.awaitEvent(AmpEvents.LOAD_START, 1).then(() => {
-        fixture.doc.body.classList.add('amp-mode-mouse');
-        const amp = fixture.doc.querySelector('#carousel-1');
-        const impl = amp.implementation_;
-
-        const prevBtn = amp.querySelector('.amp-carousel-button-prev');
-        const nextBtn = amp.querySelector('.amp-carousel-button-next');
-        prevBtn.style.visibility = 'visible';
-        nextBtn.style.visibility = 'visible';
-        expect(prevBtn).to.be.visible;
-        expect(nextBtn).to.be.visible;
-        expect(prevBtn).to.have.class('amp-disabled');
-        impl.go(-1, false);
-        expect(prevBtn).to.have.class('amp-disabled');
-        impl.go(1, false);
-        expect(prevBtn).to.not.have.class('amp-disabled');
-        impl.go(1, false);
-        impl.go(1, false);
-        expect(nextBtn).to.have.class('amp-disabled');
-        impl.go(-1, false);
-        expect(prevBtn).to.not.have.class('amp-disabled');
-        impl.go(-1, false);
-        expect(prevBtn).to.have.class('amp-disabled');
-      });
-    });
-
-    it('(type=slide) should only have the next button enabled when on ' +
-       'first item', () => {
-      return fixture.awaitEvent(AmpEvents.LOAD_START, 4).then(() => {
-        fixture.doc.body.classList.add('amp-mode-mouse');
-        const amp = fixture.doc.querySelector('#carousel-4');
-
-        const prevBtn = amp.querySelector('.amp-carousel-button-prev');
-        const nextBtn = amp.querySelector('.amp-carousel-button-next');
-        expect(prevBtn).to.be.hidden;
-        expect(nextBtn).to.be.visible;
-      });
-    });
-
-    it('should only have the prev button enabled ' +
-       'when on last item', () => {
-      return fixture.awaitEvent(AmpEvents.LOAD_START, 1).then(() => {
-        fixture.doc.body.classList.add('amp-mode-mouse');
-        const amp = fixture.doc.querySelector('#carousel-1');
-        const impl = amp.implementation_;
-        const prevBtn = amp.querySelector('.amp-carousel-button-prev');
-        const nextBtn = amp.querySelector('.amp-carousel-button-next');
-        expect(prevBtn).to.have.class('amp-disabled');
-        expect(nextBtn).to.not.have.class('amp-disabled');
-        impl.go(1, false);
-        impl.go(1, false);
-        impl.go(1, false);
-        expect(prevBtn).to.not.have.class('amp-disabled');
-        expect(nextBtn).to.have.class('amp-disabled');
-      });
-    });
-
-    it('(type=slides) should only have the prev button enabled when ' +
-       'on last item', () => {
-      return fixture.awaitEvent(AmpEvents.LOAD_START, 4).then(() => {
-        fixture.doc.body.classList.add('amp-mode-mouse');
-        const amp = fixture.doc.querySelector('#carousel-4');
-        const impl = amp.implementation_;
-        const prevBtn = amp.querySelector('.amp-carousel-button-prev');
-        const nextBtn = amp.querySelector('.amp-carousel-button-next');
-        expect(prevBtn).to.have.class('amp-disabled');
-        expect(nextBtn).to.not.have.class('amp-disabled');
-        impl.go(1, false);
-        expect(prevBtn).to.not.have.class('amp-disabled');
-        expect(nextBtn).to.have.class('amp-disabled');
-      });
-    });
-
-    it('(type=slides loop) should always have a prev and next button be ' +
-       'able to get past the first and last item', () => {
-      return fixture.awaitEvent(AmpEvents.LOAD_START, 7).then(() => {
-        fixture.doc.body.classList.add('amp-mode-mouse');
-        const amp = fixture.doc.querySelector('#carousel-7');
-        const prevBtn = amp.querySelector('.amp-carousel-button-prev');
-        const nextBtn = amp.querySelector('.amp-carousel-button-next');
-        expect(prevBtn).to.not.have.class('amp-disabled');
-        expect(nextBtn).to.not.have.class('amp-disabled');
-        nextBtn.click();
-        expect(prevBtn).to.not.have.class('amp-disabled');
-        expect(nextBtn).to.not.have.class('amp-disabled');
-        nextBtn.click();
-        expect(prevBtn).to.not.have.class('amp-disabled');
-        expect(nextBtn).to.not.have.class('amp-disabled');
-        nextBtn.click();
-        expect(prevBtn).to.not.have.class('amp-disabled');
-        expect(nextBtn).to.not.have.class('amp-disabled');
-        nextBtn.click();
-        expect(prevBtn).to.not.have.class('amp-disabled');
-        expect(nextBtn).to.not.have.class('amp-disabled');
-        nextBtn.click();
-        expect(prevBtn).to.not.have.class('amp-disabled');
-        expect(nextBtn).to.not.have.class('amp-disabled');
-      });
+    it('should not have the buttons visible ' +
+        'when amp-mode-mouse class is not on body', () => {
+      document.body.classList.remove('amp-mode-mouse');
+      return Promise.resolve().then(timeout(2000))
+          .then(() => {
+            const amp = document.querySelector('#carousel-1');
+            const prevBtn = amp.querySelector('.amp-carousel-button-prev');
+            const nextBtn = amp.querySelector('.amp-carousel-button-next');
+            expect(document.body).to.not.have.class('amp-mode-mouse');
+            expect(prevBtn).to.not.be.null;
+            expect(nextBtn).to.not.be.null;
+            expect(prevBtn).to.be.hidden;
+            expect(nextBtn).to.be.hidden;
+          });
     });
 
     it('should not have any buttons enabled when theres only a single ' +
        'item', () => {
-      return fixture.awaitEvent(AmpEvents.LOAD_START, 2).then(() => {
-        fixture.doc.body.classList.add('amp-mode-mouse');
-        const amp = fixture.doc.querySelector('#carousel-2');
+      document.body.classList.add('amp-mode-mouse');
+      return Promise.resolve().then(timeout(2000))
+          .then(() => {
+            const amp = document.querySelector('#carousel-1');
+            const prevBtn = amp.querySelector('.amp-carousel-button-prev');
+            const nextBtn = amp.querySelector('.amp-carousel-button-next');
+            expect(prevBtn).to.not.be.null;
+            expect(nextBtn).to.not.be.null;
+            expect(prevBtn).to.be.hidden;
+            expect(nextBtn).to.be.hidden;
+          });
+    });
+  });
 
-        const prevBtn = amp.querySelector('.amp-carousel-button-prev');
-        const nextBtn = amp.querySelector('.amp-carousel-button-next');
-        expect(prevBtn).to.be.hidden;
-        expect(nextBtn).to.be.hidden;
-        expect(prevBtn).to.have.class('amp-disabled');
-        expect(nextBtn).to.have.class('amp-disabled');
-      });
+  const carouselMultipleImages = `
+  <amp-carousel width=400 height=300 id="carousel-1">
+    <amp-img src="https://lh3.googleusercontent.com/pSECrJ82R7-AqeBCOEPGPM9iG9OEIQ_QXcbubWIOdkY=w400-h300-no" layout=fill></amp-img>
+    <amp-img src="https://lh3.googleusercontent.com/5rcQ32ml8E5ONp9f9-Rf78IofLb9QjS5_0mqsY1zEFc=w400-h300-no" width=400 height=300></amp-img>
+    <amp-img src="https://lh3.googleusercontent.com/Z4gtm5Bkxyv21Z2PtbTf95Clb9AE4VTR6olbBKYrenM=w400-h300-no" width=400 height=300></amp-img>
+  </amp-carousel>
+  `;
+
+  describes.integration('type=carousel with multiple images', {
+    body: carouselMultipleImages,
+    extensions,
+  }, env => {
+    let document;
+
+    beforeEach(() => {
+      document = env.win.document;
+    });
+
+    it('should be present', () => {
+      expect(document.querySelectorAll('amp-carousel')).to.have.length.above(0);
+    });
+
+    it('should not have the buttons visible ' +
+        'when amp-mode-mouse class is not on body', () => {
+      document.body.classList.remove('amp-mode-mouse');
+      return Promise.resolve().then(timeout(2000))
+          .then(() => {
+            const amp = document.querySelector('#carousel-1');
+            const prevBtn = amp.querySelector('.amp-carousel-button-prev');
+            const nextBtn = amp.querySelector('.amp-carousel-button-next');
+            expect(document.body).to.not.have.class('amp-mode-mouse');
+            expect(prevBtn).to.not.be.null;
+            expect(nextBtn).to.not.be.null;
+            expect(prevBtn).to.be.hidden;
+            expect(nextBtn).to.be.hidden;
+          });
+    });
+
+    it('should have the buttons visible when amp-mode-mouse ' +
+        'class is not on body & `controls` specified', () => {
+      document.body.classList.remove('amp-mode-mouse');
+      const amp = document.querySelector('#carousel-1');
+      amp.setAttribute('controls', '');
+      return Promise.resolve().then(timeout(2000))
+          .then(() => {
+            const prevBtn = amp.querySelector('.amp-carousel-button-prev');
+            const nextBtn = amp.querySelector('.amp-carousel-button-next');
+            expect(document.body).to.not.have.class('amp-mode-mouse');
+            expect(prevBtn).to.not.be.null;
+            expect(nextBtn).to.not.be.null;
+            expect(prevBtn).to.be.hidden;
+            expect(nextBtn).to.be.visible;
+          });
+    });
+
+    it('should only have the next button enabled ' +
+       'when on first item', () => {
+      document.body.classList.add('amp-mode-mouse');
+      return Promise.resolve().then(timeout(2000))
+          .then(() => {
+            const amp = document.querySelector('#carousel-1');
+            const prevBtn = amp.querySelector('.amp-carousel-button-prev');
+            const nextBtn = amp.querySelector('.amp-carousel-button-next');
+            expect(prevBtn).to.not.be.null;
+            expect(nextBtn).to.not.be.null;
+            expect(prevBtn).to.be.hidden;
+            expect(nextBtn).to.be.visible;
+          });
+    });
+
+    it('should not be able to go past the first or last item', () => {
+      document.body.classList.add('amp-mode-mouse');
+      return Promise.resolve().then(timeout(2000))
+          .then(() => {
+            const amp = document.querySelector('#carousel-1');
+            const impl = amp.implementation_;
+            const prevBtn = amp.querySelector('.amp-carousel-button-prev');
+            const nextBtn = amp.querySelector('.amp-carousel-button-next');
+            expect(prevBtn).to.have.class('amp-disabled');
+            impl.go(-1, false);
+            expect(prevBtn).to.have.class('amp-disabled');
+            impl.go(1, false);
+            expect(prevBtn).to.not.have.class('amp-disabled');
+            impl.go(1, false);
+            impl.go(1, false);
+            expect(nextBtn).to.have.class('amp-disabled');
+            impl.go(-1, false);
+            expect(prevBtn).to.not.have.class('amp-disabled');
+            impl.go(-1, false);
+            expect(prevBtn).to.have.class('amp-disabled');
+          });
+    });
+
+    it('should only have the prev button enabled ' +
+       'when on last item', () => {
+      document.body.classList.add('amp-mode-mouse');
+      return Promise.resolve().then(timeout(5000))
+          .then(() => {
+            const amp = document.querySelector('#carousel-1');
+            const impl = amp.implementation_;
+            const prevBtn = amp.querySelector('.amp-carousel-button-prev');
+            const nextBtn = amp.querySelector('.amp-carousel-button-next');
+            expect(prevBtn).to.have.class('amp-disabled');
+            expect(nextBtn).to.not.have.class('amp-disabled');
+            impl.go(1, false);
+            impl.go(1, false);
+            impl.go(1, false);
+            expect(prevBtn).to.not.have.class('amp-disabled');
+            expect(nextBtn).to.have.class('amp-disabled');
+          });
+    });
+  });
+
+  const slidesSingleImage = `
+  <amp-carousel width=400 height=300 type=slides id="carousel-1">
+    <amp-img src="https://lh3.googleusercontent.com/pSECrJ82R7-AqeBCOEPGPM9iG9OEIQ_QXcbubWIOdkY=w400-h300-no"></amp-img>
+  </amp-carousel>
+  `;
+
+  describes.integration('type=slides with single image', {
+    body: slidesSingleImage,
+    extensions,
+  }, env => {
+    let document;
+
+    beforeEach(() => {
+      document = env.win.document;
+    });
+
+    it('should be present', () => {
+      expect(document.querySelectorAll('amp-carousel')).to.have.length.above(0);
+    });
+
+    it('should not have the buttons visible ' +
+        'when amp-mode-mouse class is not on body', () => {
+      document.body.classList.remove('amp-mode-mouse');
+      return Promise.resolve().then(timeout(2000))
+          .then(() => {
+            const amp = document.querySelector('#carousel-1');
+            const prevBtn = amp.querySelector('.amp-carousel-button-prev');
+            const nextBtn = amp.querySelector('.amp-carousel-button-next');
+            expect(document.body).to.not.have.class('amp-mode-mouse');
+            expect(prevBtn).to.not.be.null;
+            expect(nextBtn).to.not.be.null;
+            expect(prevBtn).to.be.hidden;
+            expect(nextBtn).to.be.hidden;
+          });
     });
 
     it('(type=slides) should not have any buttons enabled when theres ' +
        'only a single item', () => {
-      return fixture.awaitEvent(AmpEvents.LOAD_START, 5).then(() => {
-        fixture.doc.body.classList.add('amp-mode-mouse');
-        const amp = fixture.doc.querySelector('#carousel-5');
-
-        const prevBtn = amp.querySelector('.amp-carousel-button-prev');
-        const nextBtn = amp.querySelector('.amp-carousel-button-next');
-        expect(prevBtn).to.be.hidden;
-        expect(nextBtn).to.be.hidden;
-        expect(prevBtn).to.have.class('amp-disabled');
-        expect(nextBtn).to.have.class('amp-disabled');
-      });
+      document.body.classList.add('amp-mode-mouse');
+      return Promise.resolve().then(timeout(2000))
+          .then(() => {
+            const amp = document.querySelector('#carousel-1');
+            const prevBtn = amp.querySelector('.amp-carousel-button-prev');
+            const nextBtn = amp.querySelector('.amp-carousel-button-next');
+            expect(prevBtn).to.not.be.null;
+            expect(nextBtn).to.not.be.null;
+            expect(prevBtn).to.be.hidden;
+            expect(nextBtn).to.be.hidden;
+          });
     });
   });
 
-  describe('when amp-mode-mouse class is not on body', () => {
+  const slidesMultipleImages = `
+  <amp-carousel width=400 height=300 type=slides id="carousel-1">
+    <amp-img src="https://lh3.googleusercontent.com/pSECrJ82R7-AqeBCOEPGPM9iG9OEIQ_QXcbubWIOdkY=w400-h300-no"></amp-img>
+    <amp-img src="https://lh3.googleusercontent.com/5rcQ32ml8E5ONp9f9-Rf78IofLb9QjS5_0mqsY1zEFc=w400-h300-no" width=400 height=300></amp-img>
+  </amp-carousel>
+  `;
 
-    it('should not have the buttons visible', () => {
-      return fixture.awaitEvent(AmpEvents.LOAD_START, 1).then(() => {
-        fixture.doc.body.classList.remove('amp-mode-mouse');
-        const amp = fixture.doc.querySelector('#carousel-1');
-        expect(fixture.doc.body).to.not.have.class('amp-mode-mouse');
+  describes.integration('type=slides with multiple images', {
+    body: slidesMultipleImages,
+    extensions,
+  }, env => {
+    let document;
 
-        const prevBtn = amp.querySelector('.amp-carousel-button-prev');
-        const nextBtn = amp.querySelector('.amp-carousel-button-next');
-        expect(prevBtn).to.be.hidden;
-        expect(nextBtn).to.be.hidden;
-      });
+    beforeEach(() => {
+      document = env.win.document;
     });
 
-    it('(type=slides) should not have the buttons visible', () => {
-      return fixture.awaitEvent(AmpEvents.LOAD_START, 4).then(() => {
-        fixture.doc.body.classList.remove('amp-mode-mouse');
-        const amp = fixture.doc.querySelector('#carousel-4');
-        expect(fixture.doc.body).to.not.have.class('amp-mode-mouse');
-
-        const prevBtn = amp.querySelector('.amp-carousel-button-prev');
-        const nextBtn = amp.querySelector('.amp-carousel-button-next');
-        expect(prevBtn).to.be.hidden;
-        expect(nextBtn).to.be.hidden;
-      });
-    });
-  });
-
-  describe('when amp-carousel has explicit `controls` attribute', () => {
-
-    it('should have visible buttons even when `amp-mode-mouse` ' +
-       'is not on body', () => {
-      return fixture.awaitEvent(AmpEvents.LOAD_START, 3).then(() => {
-        fixture.doc.body.classList.remove('amp-mode-mouse');
-        const amp = fixture.doc.querySelector('#carousel-3');
-        expect(fixture.doc.body).to.not.have.class('amp-mode-mouse');
-
-        expect(amp).to.have.attribute('controls');
-        expect(amp).to.have.class('i-amphtml-carousel-has-controls');
-
-        const prevBtn = amp.querySelector('.amp-carousel-button-prev');
-        const nextBtn = amp.querySelector('.amp-carousel-button-next');
-        expect(prevBtn).to.be.hidden;
-        expect(nextBtn).to.be.visible;
-      });
+    it('should be present', () => {
+      expect(document.querySelectorAll('amp-carousel')).to.have.length.above(0);
     });
 
-    it('(type=slides) should have visible buttons ' +
-       'even when `amp-mode-mouse` is not on body', () => {
-      return fixture.awaitEvent(AmpEvents.LOAD_START, 6).then(() => {
-        fixture.doc.body.classList.remove('amp-mode-mouse');
-        const amp = fixture.doc.querySelector('#carousel-6');
-        expect(fixture.doc.body).to.not.have.class('amp-mode-mouse');
+    it('should not have the buttons visible ' +
+        'when amp-mode-mouse class is not on body', () => {
+      document.body.classList.remove('amp-mode-mouse');
+      return Promise.resolve().then(timeout(2000))
+          .then(() => {
+            const amp = document.querySelector('#carousel-1');
+            const prevBtn = amp.querySelector('.amp-carousel-button-prev');
+            const nextBtn = amp.querySelector('.amp-carousel-button-next');
+            expect(document.body).to.not.have.class('amp-mode-mouse');
+            expect(prevBtn).to.not.be.null;
+            expect(nextBtn).to.not.be.null;
+            expect(prevBtn).to.be.hidden;
+            expect(nextBtn).to.be.hidden;
+          });
+    });
 
-        expect(amp).to.have.attribute('controls');
-        expect(amp).to.have.class('i-amphtml-carousel-has-controls');
+    it('should have the buttons visible when amp-mode-mouse ' +
+        'class is not on body & `controls` specified', () => {
+      document.body.classList.remove('amp-mode-mouse');
+      const amp = document.querySelector('#carousel-1');
+      amp.setAttribute('controls', '');
+      return Promise.resolve().then(timeout(2000))
+          .then(() => {
+            const prevBtn = amp.querySelector('.amp-carousel-button-prev');
+            const nextBtn = amp.querySelector('.amp-carousel-button-next');
+            expect(document.body).to.not.have.class('amp-mode-mouse');
+            expect(prevBtn).to.not.be.null;
+            expect(nextBtn).to.not.be.null;
+            expect(prevBtn).to.be.hidden;
+            expect(nextBtn).to.be.visible;
+          });
+    });
 
-        const prevBtn = amp.querySelector('.amp-carousel-button-prev');
-        const nextBtn = amp.querySelector('.amp-carousel-button-next');
-        expect(prevBtn).to.be.hidden;
-        expect(nextBtn).to.be.visible;
-      });
+    it('should only have the next button enabled ' +
+       'when on first item', () => {
+      document.body.classList.add('amp-mode-mouse');
+      return Promise.resolve().then(timeout(2000))
+          .then(() => {
+            const amp = document.querySelector('#carousel-1');
+            const prevBtn = amp.querySelector('.amp-carousel-button-prev');
+            const nextBtn = amp.querySelector('.amp-carousel-button-next');
+            expect(prevBtn).to.not.be.null;
+            expect(nextBtn).to.not.be.null;
+            expect(prevBtn).to.be.hidden;
+            expect(nextBtn).to.be.visible;
+          });
+    });
+
+    it('should only have the prev button enabled ' +
+       'when on last item', () => {
+      document.body.classList.add('amp-mode-mouse');
+      return Promise.resolve().then(timeout(5000))
+          .then(() => {
+            const amp = document.querySelector('#carousel-1');
+            const impl = amp.implementation_;
+            const prevBtn = amp.querySelector('.amp-carousel-button-prev');
+            const nextBtn = amp.querySelector('.amp-carousel-button-next');
+            expect(prevBtn).to.have.class('amp-disabled');
+            expect(nextBtn).to.not.have.class('amp-disabled');
+            impl.go(1, false);
+            expect(prevBtn).to.not.have.class('amp-disabled');
+            expect(nextBtn).to.have.class('amp-disabled');
+          });
+    });
+
+    it.skip('(type=slides loop) should always have a prev and next button be ' +
+    'able to get past the first and last item', () => {
+      document.body.classList.add('amp-mode-mouse');
+      const amp = document.querySelector('#carousel-1');
+      amp.setAttribute('controls', '');
+      amp.setAttribute('loop', '');
+      return Promise.resolve().then(timeout(2000))
+          .then(() => {
+            const prevBtn = amp.querySelector('.amp-carousel-button-prev');
+            const nextBtn = amp.querySelector('.amp-carousel-button-next');
+            expect(prevBtn).to.not.be.null;
+            expect(nextBtn).to.not.be.null;
+            expect(prevBtn).to.be.visible;
+            expect(nextBtn).to.be.visible;
+            expect(prevBtn).to.not.have.class('amp-disabled');
+            expect(nextBtn).to.not.have.class('amp-disabled');
+            nextBtn.click();
+            expect(prevBtn).to.not.have.class('amp-disabled');
+            expect(nextBtn).to.not.have.class('amp-disabled');
+            nextBtn.click();
+            expect(prevBtn).to.not.have.class('amp-disabled');
+            expect(nextBtn).to.not.have.class('amp-disabled');
+            nextBtn.click();
+            expect(prevBtn).to.not.have.class('amp-disabled');
+            expect(nextBtn).to.not.have.class('amp-disabled');
+            nextBtn.click();
+            expect(prevBtn).to.not.have.class('amp-disabled');
+            expect(nextBtn).to.not.have.class('amp-disabled');
+            nextBtn.click();
+            expect(prevBtn).to.not.have.class('amp-disabled');
+            expect(nextBtn).to.not.have.class('amp-disabled');
+          });
     });
   });
 });
+
+function timeout(ms) {
+  return () => new Promise(resolve => setTimeout(resolve, ms));
+}
